+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 axum = "0.6.20"
 censor = "0.3.0"
+chrono = "0.4.30"
 config = "0.13.3"
 dotenvy = "0.15.7"
 once_cell = "1.18.0"
