+++ src/utils/datetime.rs
@@ -1,0 +1,6 @@
+use chrono::Local;
+
+pub fn datetime_now() -> String {
+    let now = Local::now();
+    now.format("%v\n%r").to_string()
+}
